hack to get lines working

+++ src/commandParsing/CommandParser.java
@@ -14,7 +14,7 @@ import drawableobject.DrawableObject;
 public abstract class CommandParser {
 	
 	protected static State state;
-	protected List<Float> expressionComponents = new ArrayList<Float>();
+	protected List<Double> expressionComponents = new ArrayList<Double>();
 	
 	public void setState(State someState){
 		state = someState;
@@ -41,9 +41,9 @@ public abstract class CommandParser {
 		}
 	}
 	
-	private Float decodeStringToNumber(String stringOfInterest) throws RunTimeNullPointerException {
+	private Double decodeStringToNumber(String stringOfInterest) throws RunTimeNullPointerException {
 		if(isStringParsableAsFloat(stringOfInterest)){
-			return Float.parseFloat(stringOfInterest);
+			return Double.parseDouble(stringOfInterest);
 		}
 		else{
 			return state.fetchVariable(stringOfInterest);
