+++ src/commandParsing/mathCommandParsing/Remainder.java
@@ -12,7 +12,7 @@ import drawableobject.DrawableObject;
 public class Remainder extends TwoInputFloatCommandParser {
 
 	@Override
-	protected double operateOnComponents(List<Float> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
+	protected double operateOnComponents(List<Double> components,	Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException {
 		if(components.get(0)==0 | components.get(1)==0){
 			throw new RunTimeDivideByZeroException();
 		}
