+++ src/commandParsing/turtleCommandParsing/Backward.java
@@ -12,7 +12,7 @@ import drawableobject.DrawableObject;
 public class Backward extends OneInputFloatCommandParser {
 	
 	@Override
-	protected double operateOnComponents(List<Float> components, Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException{
+	protected double operateOnComponents(List<Double> components, Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException{
 		state.moveBackward(expressionComponents.get(0));
         objectQueue.add(new DrawableObject());
 		return expressionComponents.get(0);
