+++ src/commandParsing/turtleCommandParsing/Forward.java
@@ -13,7 +13,7 @@ import drawableobject.DrawableObject;
 public class Forward extends OneInputFloatCommandParser implements LineGenerator{
 
     @Override
-	protected double operateOnComponents(List<Float> components, Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException{
+	protected double operateOnComponents(List<Double> components, Queue<DrawableObject> objectQueue) throws RunTimeDivideByZeroException{
 		double distanceToMove = expressionComponents.get(0);
     	state.moveForward(distanceToMove);
         objectQueue.add(generateDrawableObjectRepresntingLine(distanceToMove, state));
